Refactor to provide a clear architecture and ability to extend supported formats
Split a parser to a provider and a converter, where the provider takes charge of providing headwords and corresponding content body and the converter build formatted content according to target format. Therefore, the parser deal with input and that deal with output is separated and allow a more flexible method to combine a provider and converter to meet the needs. Still, Runnable class needs to be rewrite to support working with multiple formats from CLI.